resolve email addresses

+++ app/src/main/java/org/stepik/android/adaptive/api/profile/ProfileRepositoryImpl.kt
@@ -21,19 +21,29 @@ constructor(
     override fun fetchProfile(): Single<Profile> =
             profileService.profile.map { it.profile!! }
 
+    private fun fetchProfileWithEmailAddresses(): Single<Profile> =
+            fetchProfile().flatMap { profile -> profileService.getEmailAddresses(profile.emailAddresses).map {
+                profile.emailAddressesResolved = it.emailAddresses!!
+                profile
+            }}
+
     override fun updateProfile(profile: Profile): Completable =
             profileService.setProfile(profile.id, ProfileRequest(profile))
 
     override fun updatePassword(profileId: Long, oldPassword: String, newPassword: String): Completable =
             profileService.changePassword(profileId, ChangePasswordRequest(currentPassword = oldPassword, newPassword = newPassword))
 
-    override fun updateEmail(newEmail: String): Completable = fetchProfile().flatMapCompletable { profile ->
+    override fun updateEmail(newEmail: String): Completable = fetchProfileWithEmailAddresses().flatMapCompletable { profile ->
+        if (profile.emailAddressesResolved.any { it.email == newEmail }) {
+            return@flatMapCompletable Completable.complete()
+        }
+
         val replaceAddressRx =
                 profileService.createEmailAddress(EmailAddressRequest(EmailAddress(user = profile.id, email = newEmail))).flatMapCompletable {
                     profileService.setEmailAsPrimary(it.emailAddress?.id ?: 0)
                 }
 
-        val oldEmailId = profile.emailAddresses?.firstOrNull()
+        val oldEmailId = profile.emailAddresses.firstOrNull()
         if (oldEmailId != null) {
             replaceAddressRx then profileService.removeEmailAddress(oldEmailId)
         } else {
